+++ routes/api.php
@@ -34,6 +34,7 @@ Route::middleware(['auth:web', 'throttle:api'])->group(function () {
     // Async Spotify data endpoints
     Route::get('/artists/{id}/top-tracks', [ArtistController::class, 'topTracks'])->name('api.artists.top-tracks');
     Route::get('/artists/{id}/albums', [ArtistController::class, 'albums'])->name('api.artists.albums');
+    Route::get('/artists/{id}/similar', [ArtistController::class, 'similar'])->name('api.artists.similar');
 });
 
 // User info endpoint (example)